+++ sandpack-react/src/themes/index.ts
@@ -79,7 +79,7 @@ export const sandpackDark: SandpackTheme = {
     defaultText: "#5a5a5a",
     inactiveText: "#1a1a1a",
     activeBackground: "#272727",
-    defaultBackground: "#131313",
+    defaultBackground: "#151515",
     inputBackground: "#2e2e2e",
     accent: "#90e86f",
     errorBackground: "#dac1fb",