+++ website/landing/components/AdvancedUsage/AdvancedUsage.tsx
@@ -1,7 +1,7 @@
 import { useTransform, useViewportScroll } from "framer-motion";
 import React, { useLayoutEffect, useRef, useState } from "react";
 
-// import config from "../../website.config.json";
+import { advancedUsage as content } from "../../website.config.json";
 import {
   List,
   ListItem,
@@ -15,32 +15,6 @@ import { ParallaxLogo } from "../common/ParallaxLogo";
 import { UsageExample } from "./UsageExample";
 
 export const AdvancedUsage: React.FC = () => {
-  // TODO: investigate why the app isn't able to find the existing
-  // content from the website config.
-  const content = {
-    title: "Building blocks<br />of the Sandpack<br />ecosystem.",
-    examples: [
-      {
-        title: "Providers",
-        description:
-          "We export a <code>SandpackProvider</code> that manages the state and actions for sandpack and a <code>ThemeProvider</code> that handles the theme and style.",
-        illustrationKey: "providers",
-      },
-      {
-        title: "Components",
-        description:
-          "We also export the smaller parts that make up the <code>&lt;Sandpack /&gt;</code> component, so you can combine them with your custom components.",
-        illustrationKey: "components",
-      },
-      {
-        title: "Client",
-        description:
-          "The bundler is wrapped by a small package called <code>sandpack-client</code>, which is framework agnostic and allows you to tap into the bundler protocol.",
-        illustrationKey: "client",
-      },
-    ],
-  };
-
   const section = useRef<HTMLDivElement>(null);
   const [sectionTop, setSectionTop] = useState(0);
   const [windowHeight, setWindowHeight] = useState(0);
@@ -85,19 +59,19 @@ export const AdvancedUsage: React.FC = () => {
         <List
           as="ul"
           css={{
-            display: "flex",
-            flexWrap: "wrap",
             gap: "100px",
             width: "100%",
 
             "@bp1": {
-              gap: "200px",
+              "--gap": "200px",
             },
 
             "@bp2": {
+              display: "flex",
+              flexWrap: "wrap",
               alignItems: "center",
               flexDirection: "column",
-              gap: "0",
+              "--gap": "0",
               scrollSnapType: "y mandatory",
               width: "initial",
             },