+++ website/landing/components/Intro/Sections/Template.tsx
@@ -22,6 +22,7 @@ import {
   SandpackContainerMobile,
   FadeAnimation,
   THRESHOLD_VIEW,
+  Caption,
 } from "./common";
 
 const frameworkOptions: SandpackPredefinedTemplate[] = [
@@ -73,9 +74,17 @@ export const TemplateExample: React.FC = () => {
         <Content>
           <CardTitle>Get started in a few lines</CardTitle>
           <CardDescription>
-            Set the <code>template</code> prop to get started with Sandpack in a
-            few lines of code.
+            Set the{" "}
+            <a
+              href="https://sandpack.codesandbox.io/docs/getting-started/custom-content#template"
+              rel="noreferrer"
+              target="_blank"
+            >
+              <code>template</code>
+            </a>{" "}
+            prop to get started with Sandpack in a few lines of code.
           </CardDescription>
+
           <SnippetButton
             ref={boxRef}
             onClick={shuffleTemplate}
@@ -104,7 +113,7 @@ export const TemplateExample: React.FC = () => {
                     y: mousePosition.y,
                   }}
                 >
-                  click to change
+                  Click to change
                 </ToolTip>
               </AnimatePresence>
             )}
@@ -112,6 +121,7 @@ export const TemplateExample: React.FC = () => {
         </Content>
 
         <SandpackContainerMobile>
+          <Caption>Sandpack preview</Caption>
           <SandpackPreview options={{ template }} />
         </SandpackContainerMobile>