+++ website/landing/components/Intro/Sections/common.tsx
@@ -19,29 +19,38 @@ export const Row = forwardRef<unknown, { children: React.ReactNode }>(
       <Box
         ref={ref as any}
         css={{
-          alignItems: "center",
-          display: "flex",
-          flexDirection: "column",
-          gap: "40px",
-          justifyContent: "center",
           width: "100%",
-
           "@bp2": {
-            alignItems: "center",
-            gap: "240px",
-            flexDirection: "row",
+            width: "initial",
             height: "100vh",
             maxHeight: "1080px",
-            scrollSnapAlign: "center",
-            width: "initial",
-          },
-
-          "@bp3": {
-            gap: "320px",
           },
         }}
       >
-        {children}
+        <Box
+          css={{
+            alignItems: "center",
+            display: "flex",
+            flexDirection: "column",
+            justifyContent: "center",
+            height: "100%",
+
+            gap: "40px",
+
+            "@bp2": {
+              "--gap": "240px",
+              alignItems: "center",
+              flexDirection: "row",
+              scrollSnapAlign: "center",
+            },
+
+            "@bp3": {
+              "--gap": "320px",
+            },
+          }}
+        >
+          {children}
+        </Box>
       </Box>
     );
   }
@@ -51,17 +60,16 @@ export const Content = styled("div", {
   alignItems: "center",
   display: "flex",
   flexDirection: "column",
-  gap: "20px",
+  gap: "10px",
   width: "100%",
 
   "@bp1": {
-    gap: "20px",
     width: "384px",
   },
 
   "@bp2": {
     alignItems: "flex-start",
-    width: "30%",
+    width: "450px",
   },
 });
 
@@ -163,8 +171,7 @@ export const FadeAnimation: React.FC = ({ children }) => {
   const sectionRef = useRef<HTMLLIElement>(null);
   const [sectionTop, setSectionTop] = useState(0);
   const [sectionHeight, setSectionHeight] = useState(0);
-
-  const isDesktop = useBreakpoint("bp2");
+  const shouldAnimate = useBreakpoint("bp2");
 
   useLayoutEffect(() => {
     const sectionEl = sectionRef.current;
@@ -191,7 +198,7 @@ export const FadeAnimation: React.FC = ({ children }) => {
       sectionTop + sectionHeight / 4,
       sectionTop + sectionHeight / 2,
     ],
-    [0, 1, 1, 1, 1, 0]
+    [shouldAnimate ? 0 : 1, 1, 1, 1, 1, shouldAnimate ? 0 : 1]
   );
 
   return (
@@ -204,11 +211,17 @@ export const FadeAnimation: React.FC = ({ children }) => {
 export const CodeWrapper = styled("div", {
   position: "relative",
   width: "100%",
+  marginTop: "30px !important",
 
   pre: { padding: 0 },
 
   ".sp-code-editor": {
     borderRadius: "16px",
-    padding: "0 15px",
+  },
+});
+
+export const Caption = styled("p", {
+  "@bp1": {
+    display: "none",
   },
 });