+++ website/landing/pages/index.tsx
@@ -8,7 +8,7 @@ import { Features } from "../components/Features";
 import { Footer } from "../components/Footer";
 import { Hero } from "../components/Hero";
 import { Intro } from "../components/Intro";
-// import { Showcase } from "../components/Showcase";
+import { Showcase } from "../components/Showcase";
 import { Users } from "../components/Users";
 import { styled } from "../stitches.config";
 import content from "../website.config.json";
@@ -58,10 +58,10 @@ const Home: NextPage<HomeProps> = () => {
 
       <Main>
         <Hero />
-        <Intro />
         <Features />
+        <Intro />
         <AdvancedUsage />
-        {/* <Showcase />*/}
+        <Showcase />
         <Users />
         <Banner />
         <Community />